+++ src/components/PO/index.js
@@ -1,0 +1,3 @@
+import PO from './PO';
+
+export { PO as default };
