+++ src/components/POLine/index.js
@@ -1,0 +1,3 @@
+import POLine from './POLine';
+
+export { POLine as default };
