+++ src/components/Panes/index.js
@@ -1,5 +1,3 @@
 import Panes from './Panes';
-import PO from './PO';
-import POLine from './POLine';
 
-export { Panes, PO, POLine };
+export { Panes };
